Allow user to provide previous FEN
A feature has been added to allow a user to provide the FEN string of the previous board position when running "lc2fen.py" to predict the FEN for a single image. Help messages have been updated and some comments in the code have been rephrased for grammatical and consistent purposes as well.